backend1/requirements.txt
# requirements.txt
# File chứa danh sách các thư viện Python cần thiết cho dự án

# FastAPI và các dependencies chính
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0  # MongoDB (nếu sử dụng)

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring & Logging
python-json-logger==2.0.7

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# RAG & LLM Dependencies
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Additional ML/NLP libraries
underthesea==6.7.0  # Vietnamese text processing
datasets==2.15.0    # For data handling
huggingface-hub==0.19.4  # HuggingFace model hub

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Word Document Processing
python-docx==1.1.0  # For .docx files
python-doc==0.1.0   # For .doc files (if needed)

# OCR Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Text Processing
nltk==3.8.1
spacy==3.7.2
vietnamese-nlp==0.1.0

# Vector Database
qdrant-client==1.7.0

# Async processing
asyncio==3.4.3
aiofiles==23.2.1

# Model serving
accelerate==0.25.0
bitsandbytes==0.41.3